Make computed data available via a JSON API endpoint
Render a static JSON response at /api/v1/data that returns the exports
of @/models/data, namely, {
    "data": contents of data.json,
    "builtinTypes": contents of builtins.types.json,
    "upstreamInfo": contents of meta.json,
}. The reason to return this all in a single response is so that clients
can atomically get an upstreamInfo that matches data.

This can be used by tools that wish to work with the same data as Noogle
programmatically, without needing to rerun the build process locally.