Properly support previews in GitHub Actions
We need access to the Cloudflare token to write the preview page, but we
don't want this token to be available to code controlled by the pull
request. Before this change the action used `pull_request_target` to get
has write permissions, but as a consequence, it ran on the target branch
so that it didn't run untrusted code from the pull request, so it didn't
preview the right thing.

Instead, follow the approach outlined in
https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/
by splitting the work into two workflows in the pull request case. The
first runs the build without access to secrets and stores the build
result as a GHA artifact, and the second fetches the artifact and
publishes the preview without running untrusted code.